cni-wrapper-plugin/legacynet: extract common functions
[#157103758]
